+++ assets/src/scripts/monitoring-location/components/hydrograph/selectors/layout.js
@@ -10,7 +10,6 @@ import {mediaQuery} from 'ui/utils';
 import {getCurrentParmCd} from 'ml/selectors/time-series-selector';
 
 import {getYTickDetails} from './domain';
-import {TEMPERATURE_PARAMETERS} from './time-series-data';
 
 
 export const ASPECT_RATIO = 1 / 2;
@@ -55,14 +54,14 @@ export const getLayout = memoize(kind => createSelector(
     (state) => state.ui.windowWidth,
     getYTickDetails,
     getCurrentParmCd,
-    (width, windowWidth, yTickDetails,parmCd) => {
+    (width, windowWidth, yTickDetails) => {
         const isDesktop = mediaQuery(config.USWDS_SITE_MAX_WIDTH);
         const height = kind === 'BRUSH' ? isDesktop ? BRUSH_HEIGHT : BRUSH_MOBILE_HEIGHT : width * ASPECT_RATIO;
         const margin = isDesktop ? MARGIN : MARGIN_SMALL_DEVICE;
         const tickLengths = yTickDetails.tickValues.map(v => yTickDetails.tickFormat(v).length);
         const approxLabelLength = Math.max(...tickLengths) * 10;
-        const isTemperatureParameter =
-            TEMPERATURE_PARAMETERS.celsius.concat(TEMPERATURE_PARAMETERS.fahrenheit).includes(parmCd);
+        const hasRightYAxis = false; // placeholder for changes upcoming in ticket WDFN-370
+
         return {
             width: width,
             height: height,
@@ -71,7 +70,7 @@ export const getLayout = memoize(kind => createSelector(
                 bottom: margin.bottom,
                 top: kind === 'BRUSH' ? 13 : margin.top,
                 left: margin.left + approxLabelLength,
-                right: margin.right + (isTemperatureParameter ? approxLabelLength : 0)
+                right: margin.right + (hasRightYAxis ? approxLabelLength : 0)
             }
         };
     }