+++ assets/src/scripts/monitoring-location/components/hydrograph/selectors/parameter-data.js
@@ -1,7 +1,6 @@
 import {createSelector} from 'reselect';
 
 import {sortedParameters} from 'ui/utils';
-
 import {getCurrentVariableID, getTimeSeries, getVariables} from 'ml/selectors/time-series-selector';
 
 /**