+++ assets/src/scripts/monitoring-location/store/statistics-data.js
@@ -1,3 +1,7 @@
+import cloneDeep from 'lodash/cloneDeep';
+
+import config from 'ui/config';
+import {convertCelsiusToFahrenheit} from 'ui/utils';
 import {fetchSiteStatistics} from 'ui/web-services/statistics-data';
 
 const INITIAL_DATA = {
@@ -17,6 +21,21 @@ const setMedianStats = function(statisticsData) {
 };
 
 /*
+*  Helper function that takes a group of median value Celsius temperature statistics and converts to Fahrenheit.
+* @param {Object} clonedStatGroup - An object grouped by method IDs that contains a cloned copy of temperature value statistics
+* @return {Object} clonedStatGroup - The median statistics with converted temperatures
+*/
+const convertMethodStatsFromCelsiusToFahrenheit = function(clonedStatGroup) {
+        Object.entries(clonedStatGroup).forEach(statEntry => {
+            statEntry[1].forEach(detail => {
+                detail.p50_va = convertCelsiusToFahrenheit(detail.p50_va);
+            });
+        });
+
+    return clonedStatGroup;
+};
+
+/*
  * Asynchronous Redux action to fetch the all median statistics data for a site
  * @param {String} siteno
  * @return {Function} which returns a promise once the data has been fetched
@@ -24,6 +43,19 @@ const setMedianStats = function(statisticsData) {
 const retrieveMedianStatistics = function(siteno) {
     return function(dispatch) {
         return fetchSiteStatistics({siteno, statType: 'median'}).then((stats) => {
+            Object.entries(stats).forEach(parameterStatGroup => {
+                const parameterCode = parameterStatGroup[0];
+                const methodGroupings = parameterStatGroup[1];
+                if (config.TEMPERATURE_PARAMETERS.celsius.includes(parameterCode)) {
+                    const convertedStatGroup = convertMethodStatsFromCelsiusToFahrenheit(cloneDeep(methodGroupings));
+
+                    stats = {
+                        ...stats,
+                        [`${parameterCode}${config.CALCULATED_TEMPERATURE_VARIABLE_CODE}`]: convertedStatGroup
+                    };
+                }
+            });
+
             dispatch(setMedianStats(stats));
         });
     };